Fixed error on get_next_line + we can use only uint32_t for face indexes but it should be enough, 4 billion vertexes would melt any gpu anyway

+++ include/scop.h
@@ -27,7 +27,7 @@ typedef struct	s_vec3 {
 }				t_vec3;
 
 typedef struct	s_triangle {
-	uint64_t	points[3];
+	uint32_t	points[3];
 }				t_triangle;
 
 typedef struct	s_object {
